Closer to complete
In class today I made a change to the Ongoing type to allow for a message to be output, similar to GameOver. I added the messages throughout Lib. I added a testing initial state to Spec to allow for a control test when using the game. Lib no longer uses any IO types also. I refined the code so that instead of output it just passes the message back to IO. I also added helper functions to aid in the random seed generation. Now all hazards have their own unique spawns too.